Organizing GAN utils
